changed type of fields in schema to schow decimal numbers, commented caching

+++ src/main/resources/schema.sql
@@ -15,7 +15,7 @@ CREATE TABLE material_grade_dic(
     material_grade_id BIGINT PRIMARY KEY AUTO_INCREMENT,
     eu_symbol VARCHAR(100),
     ger_symbol VARCHAR(100),
-    density INT,
+    density DECIMAL(100000, 2),
     grade_group VARCHAR(100)
 );
 
@@ -29,10 +29,10 @@ CREATE TABLE plate_material(
     dimension_A INT,
     dimension_B INT,
     thickness INT,
-    weight INT,
+    weight DECIMAL(100000, 2),
     is_painted BOOLEAN,
     is_painted_both_sides BOOLEAN,
-    surface_to_conserve INT,
+    surface_to_conserve DECIMAL(1000, 2),
     project_id BIGINT,
     material_grade_id BIGINT
 );
